# inflate: Limit variable shifts