@klauspost 2020-08-18
Could be refined, but better for now.